# dehaze-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "去雾系统后端 API 接口文档",
        "title": "Dehaze System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@dehaze.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/dataset": {
            "get": {
                "description": "获取数据集树形列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集接口"
                ],
                "summary": "数据集树形列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keywords",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/vo.DatasetVO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "新增数据集",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集接口"
                ],
                "summary": "新增数据集",
                "parameters": [
                    {
                        "description": "数据集表单数据",
                        "name": "datasetForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bo.DatasetFormBO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.DatasetFormBO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据集",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集接口"
                ],
                "summary": "删除数据集",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据集ID，多个以英文逗号(,)分割",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataset/item": {
            "put": {
                "description": "修改数据项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集项接口"
                ],
                "summary": "修改数据项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "数据项ID",
                        "name": "datasetItemId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增数据项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集项接口"
                ],
                "summary": "新增数据项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "所属数据集ID",
                        "name": "datasetId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer",
                                            "format": "int64"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集项接口"
                ],
                "summary": "删除数据项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "数据项ID",
                        "name": "datasetItemId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataset/operations/items": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集操作"
                ],
                "summary": "创建数据项并上传配对图片",
                "parameters": [
                    {
                        "description": "创建请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/dataset/operations/items/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集操作"
                ],
                "summary": "批量创建数据项并上传配对图片",
                "parameters": [
                    {
                        "description": "批量创建请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/dataset/operations/items/{itemId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集操作"
                ],
                "summary": "级联删除数据项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "数据项ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataset/{id}": {
            "put": {
                "description": "修改数据集",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据集接口"
                ],
                "summary": "修改数据集",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "数据集ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "数据集表单数据",
                        "name": "datasetForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bo.DatasetFormBO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/files": {
            "post": {
                "description": "文件上传",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件接口"
                ],
                "summary": "文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "表单文件对象",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "模型id",
                        "name": "modelId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SysFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "文件删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件接口"
                ],
                "summary": "文件删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "fileId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bo.DatasetFormBO": {
            "type": "object",
            "required": [
                "parentId"
            ],
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "数据集描述",
                    "type": "string"
                },
                "id": {
                    "description": "数据集ID",
                    "type": "integer"
                },
                "name": {
                    "description": "数据集名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父数据集ID",
                    "type": "integer"
                },
                "path": {
                    "description": "数据集存储路径",
                    "type": "string"
                },
                "statistics": {
                    "description": "统计信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/bo.StatisticsBO"
                        }
                    ]
                },
                "status": {
                    "description": "状态(1:正常;0:禁用) - 支持字符串和数字类型",
                    "type": "integer"
                },
                "type": {
                    "description": "数据集类型",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "bo.StatisticsBO": {
            "type": "object",
            "properties": {
                "clearCount": {
                    "type": "integer"
                },
                "fileCount": {
                    "type": "integer"
                },
                "formatDistribution": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "hazeDistribution": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "hazyCount": {
                    "type": "integer"
                },
                "itemCount": {
                    "type": "integer"
                },
                "sceneDistribution": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "totalSize": {
                    "type": "integer"
                }
            }
        },
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.SysFile": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "md5": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objectName": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "vo.DatasetVO": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子数据集",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.DatasetVO"
                    }
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "数据集描述",
                    "type": "string"
                },
                "id": {
                    "description": "数据集ID",
                    "type": "integer"
                },
                "name": {
                    "description": "数据集名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父数据集ID",
                    "type": "integer"
                },
                "path": {
                    "description": "存储位置",
                    "type": "string"
                },
                "size": {
                    "description": "占用空间大小",
                    "type": "string"
                },
                "status": {
                    "description": "状态(1:启用；0:禁用)",
                    "type": "integer"
                },
                "type": {
                    "description": "数据集类型",
                    "type": "string"
                },
                "updateTime": {
                    "description": "修改时间",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}